requirements.txt
oracledb==2.4.1
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
